Rename feature to command-palette-navigation-shortcuts
