+++ source/refined-github.ts
@@ -215,4 +215,4 @@ import './features/prevent-pr-merge-panel-opening';
 import './features/dim-visited-conversations';
 import './features/rgh-improve-new-issue-form';
 import './features/easy-toggle-commit-messages';
-import './features/command-palette-vim-keybindings';
+import './features/command-palette-navigation-shortcuts';
